fixes to unit picker
